Create mian.with_new_coffeemachine
Here is the latest simulator code. However, I'm not sure whether devices like the microwave oven and media player need to interact with the server. For instance, after the countdown function of the microwave oven ends, should it send an "off" state to the server to modify the backend database state? Currently, no interaction feature has been added. The coffee machine currently operates on the principle of automatically stopping after brewing for 10 seconds, which doesn't require interaction with the server. The remaining modes, types, and countdown functions have not been tested, only the code logic. However, every on/off function of each simulator has been tested.